3. fixasi error

+++ src/api/notes/handler.js
@@ -1,45 +1,36 @@
+/* eslint-disable no-underscore-dangle */
 const ClientError = require('../../exceptions/ClientError');
 
-/* eslint-disable no-underscore-dangle */
 class NotesHandler {
   constructor(service, validator) {
     this._service = service;
     this._validator = validator;
 
     this.postNoteHandler = this.postNoteHandler.bind(this);
-
     this.getNotesHandler = this.getNotesHandler.bind(this);
-
     this.getNoteByIdHandler = this.getNoteByIdHandler.bind(this);
-
     this.putNoteByIdHandler = this.putNoteByIdHandler.bind(this);
-
     this.deleteNoteByIdHandler = this.deleteNoteByIdHandler.bind(this);
   }
 
-  postNoteHandler(request, h) {
+  async postNoteHandler(request, h) {
     try {
       this._validator.validateNotePayload(request.payload);
       const { title = 'untitled', body, tags } = request.payload;
+      const { id: credentialId } = request.auth.credentials;
 
-      const noteId = this._service.addNote({ title, body, tags });
-
-      const response = h.response({
-
-        status: 'success',
-
-        message: 'Catatan berhasil ditambahkan',
-
-        data: {
-
-          noteId,
-
-        },
-
+      const noteId = await this._service.addNote({
+        title, body, tags, owner: credentialId,
       });
 
+      const response = h.response({
+        status: 'success',
+        message: 'Catatan berhasil ditambahkan',
+        data: {
+          noteId,
+        },
+      });
       response.code(201);
-
       return response;
     } catch (error) {
       if (error instanceof ClientError) {
@@ -47,72 +38,14 @@ class NotesHandler {
           status: 'fail',
           message: error.message,
         });
-
         response.code(error.statusCode);
         return response;
       }
 
-      // SERVER ERROR!
+      // Server ERROR!
       const response = h.response({
         status: 'error',
         message: 'Maaf, terjadi kegagalan pada server kami.',
-      });
-      response.code(500);
-      console.log(error);
-      return response;
-    }
-  }
-
-  getNotesHandler() {
-    const notes = this._service.getNotes();
-
-    return {
-
-      status: 'success',
-
-      data: {
-        notes,
-      },
-
-    };
-  }
-
-  getNoteByIdHandler(request, h) {
-    try {
-      const { id } = request.params;
-
-      const note = this._service.getNoteById(id);
-
-      return {
-
-        status: 'success',
-
-        data: {
-
-          note,
-
-        },
-
-      };
-    } catch (error) {
-      if (error instanceof ClientError) {
-        const response = h.response({
-
-          status: 'fail',
-
-          message: error.message,
-
-        });
-
-        response.code(error.statusCode);
-
-        return response;
-      }
-
-      // SERVER ERROR!
-      const response = h.response({
-        status: 'error',
-        message: 'Maaf, terjadi kesalahan pada server kami',
       });
       response.code(500);
       console.error(error);
@@ -120,80 +53,112 @@ class NotesHandler {
     }
   }
 
-  putNoteByIdHandler(request, h) {
+  async getNotesHandler(request) {
+    const { id: credentialId } = request.auth.credentials;
+    const notes = await this._service.getNotes(credentialId);
+    return {
+      status: 'success',
+      data: {
+        notes,
+      },
+    };
+  }
+
+  async getNoteByIdHandler(request, h) {
     try {
-      this._validator.validateNotePayload(request.payload);
       const { id } = request.params;
+      const { id: credentialId } = request.auth.credentials;
 
-      this._service.editNoteById(id, request.payload);
-
+      await this._service.verifyNoteAccess(id, credentialId);
+      const note = await this._service.getNoteById(id);
       return {
-
         status: 'success',
-
-        message: 'Catatan berhasil diperbaharui',
-
+        data: {
+          note,
+        },
       };
     } catch (error) {
       if (error instanceof ClientError) {
         const response = h.response({
-
           status: 'fail',
-
           message: error.message,
-
         });
-
         response.code(error.statusCode);
-
         return response;
       }
 
-      // SERVER ERROR!
+      // Server ERROR!
       const response = h.response({
         status: 'error',
-        message: 'Maaf, terjadi kegagalan pada server kami',
+        message: 'Maaf, terjadi kegagalan pada server kami.',
       });
       response.code(500);
-      console.log(error);
+      console.error(error);
       return response;
     }
   }
 
-  deleteNoteByIdHandler(request, h) {
+  async putNoteByIdHandler(request, h) {
     try {
-      const { id } = request.params;
+      this._validator.validateNotePayload(request.payload);
+      const { id } = request.params; const { id: credentialId } = request.auth.credentials;
 
-      this._service.deleteNoteById(id);
-
+      await this._service.verifyNoteAccess(id, credentialId);
+      await this._service.editNoteById(id, request.payload);
       return {
-
         status: 'success',
-
-        message: 'Catatan berhasil diperbaharui',
-
+        message: 'Catatan berhasil diperbarui',
       };
     } catch (error) {
       if (error instanceof ClientError) {
         const response = h.response({
-
           status: 'fail',
-
-          message: 'Catatan gagal dihapus , Id tidak ditemukan',
-
+          message: error.message,
         });
-
         response.code(error.statusCode);
         return response;
       }
 
-      // SERVER ERROR!
+      // Server ERROR!
       const response = h.response({
         status: 'error',
-        message: 'Maaf, terjadi kegagaln pada server kami',
+        message: 'Maaf, terjadi kegagalan pada server kami.',
       });
       response.code(500);
-      console.log(error);
+      console.error(error);
+      return response;
+    }
+  }
+
+  async deleteNoteByIdHandler(request, h) {
+    try {
+      const { id } = request.params;
+      const { id: credentialId } = request.auth.credentials;
+
+      await this._service.verifyNoteOwner(id, credentialId);
+      await this._service.deleteNoteById(id);
+
+      return {
+        status: 'success',
+        message: 'Catatan berhasil dihapus',
+      };
+    } catch (error) {
+      if (error instanceof ClientError) {
+        const response = h.response({
+          status: 'fail',
+          message: error.message,
+        });
+        response.code(error.statusCode);
+        return response;
+      }
+
+      // Server ERROR!
+      const response = h.response({
+        status: 'error',
+        message: 'Maaf, terjadi kegagalan pada server kami.',
+      });
+      response.code(500);
+      console.error(error);
       return response;
     }
   }
